Crash when using an SVG font with > 390 glyphs https://bugs.webkit.org/show_bug.cgi?id=149677 <rdar://problem/21676402>
Reviewed by Simon Fraser.

Source/WebCore:

The "Charset Index" in OTF are indices into a collection of strings. There are
390 predefined strings in this collection. We were currently assigning each
glyph to one of these strings. However, if there are more glyphs than strings,
we will be using invalid indices.

The values of the strings themselves are not necessary for SVG fonts. Therefore,
the solution is to create a single dummy string, and have all glyphs target it.

Tests: svg/custom/many-glyphs.svg

* css/CSSFontFaceSource.cpp:
(WebCore::CSSFontFaceSource::font):
* svg/SVGToOTFFontConversion.cpp:
(WebCore::SVGToOTFFontConverter::appendCFFTable):

LayoutTests:

* svg/custom/many-glyphs-expected.svg: Added.
* svg/custom/many-glyphs.svg: Added.

Identifier: 167791@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190375 268f45cc-cd09-0410-ab3c-d52691b4dbfc