[Win] Unreviewed test fix.
You cannot run 'run-webkit-tests' from the command line because it does not know
to append the binary subdirectory (bin32/bin64). We don't see this on our builders
because they explicitly set use the 'root' argument to provide the full path to
the test binaries.

* Scripts/webkitpy/port/win.py:
(WinPort._build_path): Added.



Identifier: 167867@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190511 268f45cc-cd09-0410-ab3c-d52691b4dbfc