Reference cycles during SVG dependency invalidation https://bugs.webkit.org/show_bug.cgi?id=149824 <rdar://problem/22771412>
Reviewed by Tim Horton.

Source/WebCore:

Detect any reference cycles as we are invalidating.

This is mostly a merge of the following Blink commit:
https://chromium.googlesource.com/chromium/blink/+/a4bc83453bda89823b672877dc02247652a02d51

Test: svg/custom/reference-cycle.svg

* rendering/svg/RenderSVGResource.cpp:
(WebCore::removeFromCacheAndInvalidateDependencies): Keep around a hash
table of dependencies, so that we can detect if an element is already
present before marking it.

LayoutTests:

Adding a test that has a cycle between feImage resources.

Merge Blink commit:
https://chromium.googlesource.com/chromium/blink/+/a4bc83453bda89823b672877dc02247652a02d51

* svg/custom/reference-cycle-expected.txt: Added.
* svg/custom/reference-cycle.svg: Added.

Identifier: 167940@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190592 268f45cc-cd09-0410-ab3c-d52691b4dbfc