Web Inspector: Clicking on a stack trace link does not navigate to the corresponding line https://bugs.webkit.org/show_bug.cgi?id=149405
lineNumber should be a number type but it was a string.

Reviewed by Timothy Hatcher.

* UserInterface/Base/Main.js:
(WebInspector.openURL):
(WebInspector.linkifyStringAsFragmentWithCustomLinkifier):
Keep lineNumber undefined when it isn't present in the string.
WebInspector.linkifyStringAsFragmentWithCustomLinkifier is only used by
WebInspector.linkifyStringAsFragment and it expects lineNumber to be undefined
when it wasn't found in the parsed string.


Identifier: 167996@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190651 268f45cc-cd09-0410-ab3c-d52691b4dbfc