Backgrounds bleed out of natively rendered text fields https://bugs.webkit.org/show_bug.cgi?id=149843 <rdar://problem/22896977>
Reviewed by Darin Adler.

Source/WebCore:

When natively rendering a text field with a background on Mac, the background bleeds out
of the text field's border when the graphics context is scaled (as a result of a retina
display or zoom/scale effects). This is because when we render the text field in bezeled
style within a certain frame, AppKit adds 1 device pixel insets on all sides of the frame,
which renders a text field that is slightly smaller than the frame. To adjust for this, we
inflate the paint rect.

Test: fast/forms/hidpi-textfield-background-bleeding.html

* rendering/RenderThemeMac.mm:
(WebCore::RenderThemeMac::paintTextField):

LayoutTests:

Add a layout test to check that the background of a natively rendered text field
cell does not bleed.

* TestExpectations:
* fast/forms/hidpi-textfield-background-bleeding-expected.html: Added.
* fast/forms/hidpi-textfield-background-bleeding.html: Added.

Identifier: 168102@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190800 268f45cc-cd09-0410-ab3c-d52691b4dbfc