Web Inspector: Don't shadow global Object constructor in WebInspector.Object class definition https://bugs.webkit.org/show_bug.cgi?id=150093
Reviewed by Joseph Pecoraro.

Currently, class expression does not define any variables. So even in `class Object { }`,
we were able to touch the global Object constructor (like `Object.keys`).
However, that is a bug. After https://bugs.webkit.org/show_bug.cgi?id=150089 fixes that bug,
the global `Object` is shadowed by the user-defined `class Object` name.

To solve this issue, we changed the `class Object` to `class WebInspectorObject`.
And we expose this class as a `WebInspector.Object`.

* UserInterface/Base/Object.js:

Identifier: 168229@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191001 268f45cc-cd09-0410-ab3c-d52691b4dbfc