ASSERTION FAILED: computeMainAxisExtentForChild(child, MainOrPreferredSize, mainSize) in WebCore::RenderFlexibleBox::adjustChildSizeForMinAndMax https://bugs.webkit.org/show_bug.cgi?id=149459
Reviewed by Darin Adler.

Source/WebCore:

This was regressed after 189567 where min-height|width:auto
support was added to flex items. The merge from Blink changes
was not correctly done for assertions. In particular we were
asserting if the resolved main size was not strictly greater
than 0, but 0 is actually a valid value.

Test: fast/flexbox/crash-resolved-main-size-zero.html

* rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::adjustChildSizeForMinAndMax):

LayoutTests:

* fast/flexbox/crash-resolved-main-size-zero-expected.txt: Added.
* fast/flexbox/crash-resolved-main-size-zero.html: Added.

Identifier: 168503@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@191336 268f45cc-cd09-0410-ab3c-d52691b4dbfc